==> Analysis/07-compare-PCAs.Rmd <==
---
title: "04-compare_PCAs"
author: "Ariel Marcy & Vera Weisbecker"
date: "8/15/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=FALSE, message=FALSE, results='hide')

```

# Plot PCAs for allometric and shape residual datasets
Plot PCAs for allometric shape and shape residual, provide coloration and pch symbols by locomotion and diet, and visualize changes in shape residual PCA when *Notomys* species are removed. These plots are relevant to the heatplots created in script 05 / shown in Figure 2. 

Creates multipanel Figure 1.

Compares morphospace similarity of the two datasets with Mantel tests and a correlation test.

Creates supplementary Figure 1, a screeplot of the the PCs for each dataset.

Produces shape residual data used in later analyses, stored in rda file 04.

### Load packages, functions, and data
Tree, mean shape, and metadata for phylogenetic analyses come from the allometry rodents script 04. 
```{r message = FALSE}
library(geomorph)   # CRAN v4.0.7   
library(vegan)      # CRAN v2.6-4      
library(dplyr)      # CRAN v1.1.4      
library(colorspace) # CRAN v2.1-0 

# This sets the wd to local
library(rstudioapi) # Safely Access the RStudio API
setwd(dirname(getActiveDocumentContext()$path))

source("../Functions/utilities.R")  # custom functions
load(file = "../Data/Processed/03-main-data.rda")
load(file = "../Data/Processed/03-tree-data.rda")
load(file = "../Data/Processed/03-graphing-vectors.rda")
#This only contains the residuals of a pgls testing for allometry, see file 04 for why we are happy with BM models here.
load(file = "../Data/Processed/allometry_residuals.rda")
```

# 1) Export Figure 1
Create complex, multipanel plot of PCAs from different datasets

## A) PCA of mean shapes in full shape dataset (with allometry)
```{r}
mean.PCA <- gm.prcomp(mean.shapes)

# Make axis labels
PCs.allo <- mean.PCA$pc.summary$importance
PCx.per.allo <- round(PCs.allo[2, 1] * 100, digits = 1)  # % with 1 decimal
PCx.lab.allo <- paste("PC1", " (", PCx.per.allo, "%)", sep = "")
PCy.per.allo <- round(PCs.allo[2, 2] * 100, digits = 1)
PCy.lab.allo <- paste("PC2", " (", PCy.per.allo, "%)", sep = "")
```

## B-D) PCAs of shape residual dataset (WITHOUT allometry)
All other PCA plots in figure use this dataset


### Make PCA of shape residuals
Color and legend by species
```{r}
# PCA in part B and D
res.PCA <- gm.prcomp(residuals)

# Make axis labels
PCs <- res.PCA$pc.summary$importance
PCx.per <- round(PCs[2, 1] * 100, digits = 1)  # % with 1 decimal
PCx.lab <- paste("PC1", " (", PCx.per, "%)", sep = "")
PCy.per <- round(PCs[2, 2] * 100, digits = 1)
PCy.lab <- paste("PC2", " (", PCy.per, "%)", sep = "")
```

### C) Plot colors by diet and point shapes by locomotion
```{r}
# Set up palette for 5 diet types
warm <- sequential_hcl(5, "LaJolla")  
col.carn <- warm[4]  # meat colored red
col.frug <- warm[3]  # fruit colored orange
col.gran <- warm[2]  # grain colored gold
col.foli <- sequential_hcl(5, "Greens 3")[2]  # green grass color for folivores
col.omni <- sequential_hcl(5, "Grays")[3]  # gray for anything goes omnivores

# Define color palette for diets of mean specimens
col.diet.5 <- c(col.carn, col.frug, col.gran, col.foli, col.omni)
col.diet.means <- PlotByGroup(info.means, "Feeding", col.diet.5)

# Points for 4 locomotion types: Aquatic (circle), Arboreal (triangle), Terrestrial (square), hopping (diamond)
pch.gld.4 <- c(16, 24, 23, 22)
pch.gld.means <- PlotByGroup(info.means, "Guild", pch.gld.4)

# Genera legend 
genera.phylo.full <- c("Mus", "Rattus", "Pogonomys", "Xeromys", "Hydromys", "Notomys", "Pseudomys", "Zyzomys", "Leggadina", "Uromys", "Melomys", "Leporillus", "Mesembriomys", "Conilurus")
```

### D) Plot PCA of shape residuals without Notomys
The bipedal hopping locomotion of Notomys could be driving a signal not found in the remaining dataset. To test, we take these 4 species out. 
```{r}

#check check
dimnames(residuals)[[3]]==info.means$FullName
# Remove Notomys species from dataset
index.not <- which(info.means$Genus == "Not")
residuals.no <- residuals[, , -index.not]
info.means.no <- info.means[-index.not, ]

# PCA
wrap.PCA.r.n <- gm.prcomp(residuals.no)
alpha.col.no <- alpha.col[-index.not]
alpha.pch.no <- alpha.pch[-index.not]
col.diet.means.no <- col.diet.means[-index.not]
pch.gld.means.no <- pch.gld.means[-index.not]



# Calulate % of PC
PCs.n <- wrap.PCA.r.n$pc.summary$importance
PCx.per.n <- round(PCs.n[2, 1] * 100, digits = 1)  # % with 1 decimal
PCx.lab.n <- paste("PC1", " (", PCx.per.n, "%)", sep = "")
PCy.per.n <- round(PCs.n[2, 2] * 100, digits = 1)
PCy.lab.n <- paste("PC2", " (", PCy.per.n, "%)", sep = "")
```

### Prepare evolutionary allometry information (from AmNat paper) for Fig 1
```{r}
# Run Procrustes ANOVA with interaction
pgls.gdf <- geomorph.data.frame(coords = mean.shapes, size = info.means$MeanCsize, clade = info.means$Clade)

fit.pgls <- procD.pgls(coords ~ log(size) * clade, aus.tree, iter = 500, data = pgls.gdf, print.progress = FALSE)

# Find Reg Score values for plotting purposes
evo.allo.rs <- plotAllometry(fit.pgls, size = info.means$MeanCsize, method = "RegScore")
```

### Export Figure 1 to pdf
Finally, we compare the full shape PCA to the shape residuals PCAs. All PCAs are relevant to Figure 2 heatmaps.
```{r}
# Set up Notomys abline
not.index <- which(info.means$Genus.y == "Notomys")
not.Reg <- evo.allo.rs$RegScore[not.index]
not.Cs <- info.means$MeanCsize[not.index]

#Set up PCA summaries to extract importance
summary_meanPCA<- round( summary(mean.PCA)$PC.summary[2,],3)*100
summary_meanPCA_res <- round(summary(res.PCA)$PC.summary[2,],3)*100
summary_MeanPCA_res_noNoto <- round(summary(wrap.PCA.r.n)$PC.summary[2,],3)*100


# Set up for export and for multiple panels
part.coords <- c(-0.1, 0.06)  # plotting parameters for A, C, D, E placement
title_coords <- c(-0.02, 0.06)
xlims <- c(-0.10, 0.115)
ylims <- c(-0.09, 0.06)


pdf(file="../Data/Results/PCA_4_panel.pdf", height= 7.15, width = 8.64 )
mat <- matrix(c(1:4), 2, 2, byrow = TRUE)  # 4 plots, 2 rows, 2 columns
layout(mat, widths = rep(1, dim(mat)[2]), heights = rep(0.5, dim(mat)[1]))   

# A) FULL SHAPE PCA - diet & locomotion legend
par(mar = c(4.5, 4.5, 2, 1))
plot(x = -mean.PCA$x[, 1],  # Flip for consistency
                 y = mean.PCA$x[, 2],  
                 xlab = paste("PC1 (",summary_meanPCA[1],"%)", sep=""), 
                 ylab = paste("PC2 (",summary_meanPCA[2],"%)", sep=""), 
                 xlim = xlims,
                 ylim = ylims,
                 col = col.diet.means, 
                 pch = pch.gld.means, 
                 bg = col.diet.means,
                 cex = 1.5,
                 cex.axis = 1.1, 
                 cex.lab = 1.3, 
                 asp=TRUE)

text(part.coords[1], part.coords[2], "a", cex = 1.5)
text(title_coords[1],title_coords[2], "full shape dataset (ecology)")
text(-0.05, 0.05, expression(paste("Marcy ", italic("et al. "), "2020")), cex = 0.8)





# B) Evolutionary Allometry plot - diet & locomotion legend
par(mar = c(4, 4, 1, 1))
plot(x = log(info.means$MeanCsize),
     y = -evo.allo.rs$RegScore,
     xlim = c(4.75, 6),
     col = col.diet.means, 
     pch = pch.gld.means, 
     bg = col.diet.means,
     main = " ",
     xlab = "Log centroid size", 
     ylab = "Shape  (Reg Score)",
     bty = "o",
     cex.axis = 1.1,
     cex.lab = 1.3)

text(4.8, 0.0014, "b", cex = 1.5)
text(5.1, 0.00139, "evolutionary allometry")
abline(lm(-evo.allo.rs$RegScore ~ log(info.means$MeanCsize)), col = "dark grey", lwd = 1)  # common evolutionary allometry line



# C) shape Residuals PCA - ecology
par(mar = c(4.5, 4.5, 2, 1))
plot(x = res.PCA$x[, 1],  # no flip for consistency
                 y = -res.PCA$x[, 2],  
                 xlab = paste("PC1 (",summary_meanPCA_res[1],"%)", sep=""), 
                 ylab = paste("PC2 (",summary_meanPCA_res[2],"%)", sep=""), 
                 xlim = xlims,
                 ylim = ylims,
                 col = col.diet.means, 
                 pch = pch.gld.means, 
                 bg = col.diet.means,
                 cex = 1.5,
                 cex.axis = 1.1, 
                 cex.lab = 1.3, 
                 asp=TRUE)
text(part.coords[1], part.coords[2], "c", cex = 1.5)
text(title_coords[1],title_coords[2], "shape residual (ecology)")

text(res.PCA$x[which(rownames(res.PCA$x)=="Mastacomys_fuscus"),1],-res.PCA$x[which(rownames(res.PCA$x)=="Mastacomys_fuscus"),2], expression(italic( "M. fuscus")), pos=1)

text(res.PCA$x[which(rownames(res.PCA$x)=="Conilurus_penicillatus"),1],-res.PCA$x[which(rownames(res.PCA$x)=="Conilurus_penicillatus"),2], expression(italic( "C. penicill.")), pos=3)

text(res.PCA$x[which(rownames(res.PCA$x)=="Notomys_mitchellii"),1],-res.PCA$x[which(rownames(res.PCA$x)=="Notomys_mitchellii"),2], expression(italic( "Notomys")), pos=3)

legend(-0.1, -0.0, legend = c("Arboreal", "Hopping", "Terrestrial", "Semiaquatic", "", "Carnivore", "Frugivore", "Granivore", "Folivore", "Omnivore"), col = c(rep("black", 4), "white", col.diet.5), border = NULL, pch = c(24:21, rep(16, 6)), cex = 0.8, bty = "n")

# D) No Notomys shape residuals PCA - diet colors
par(mar = c(4.5, 4.5, 2, 1))
plot(x = wrap.PCA.r.n$x[, 1],  
                 y = wrap.PCA.r.n$x[, 2],  # no flip for consistency
                xlab = paste("PC1 (",summary_MeanPCA_res_noNoto[1],"%)", sep=""), 
                 ylab = paste("PC2 (",summary_MeanPCA_res_noNoto[2],"%)", sep=""),
                 xlim = xlims,
                 ylim = ylims,
                 col = col.diet.means.no, 
                 pch = pch.gld.means.no, 
                 bg = col.diet.means.no,
                 cex = 1.5,
                 cex.axis = 1.1, 
                 cex.lab = 1.3, 
                 asp=TRUE)
text(part.coords[1], part.coords[2], "d", cex = 1.5)
text(title_coords[1],title_coords[2], expression(paste("shape residual without ", italic("Notomys"))))

dev.off()
```

#Create supplementary figure for lm-based allometry residuals
```{r}
# PCA in part B and D
res_lm_PCA <- gm.prcomp(residuals_lm)

#Set up PCA summaries to extract importance
summary_meanPCA_lm_res <- round(summary(res_lm_PCA)$PC.summary[2,],3)*100


pdf(file="../Data/Results/SuppPCA_lm_residuals.pdf", height= 7.15, width = 8.64 )



#Color and legend by species


# Make axis labels
# PCs <- res_lm_PCA$pc.summary$importance
# PCx.per <- round(PCs[2, 1] * 100, digits = 1)  # % with 1 decimal
# PCx.lab <- paste("PC1", " (", PCx.per, "%)", sep = "")
# PCy.per <- round(PCs[2, 2] * 100, digits = 1)
# PCy.lab <- paste("PC2", " (", PCy.per, "%)", sep = "")
# 



# A) shape Residuals PCA - ecology
par(mar = c(4.5, 4.5, 2, 1))
plot(x = res_lm_PCA$x[, 1],  # no flip for consistency
                 y = -res_lm_PCA$x[, 2],  
                 xlab = paste("PC1 (",summary_meanPCA_lm_res[1],"%)", sep=""), 
                 ylab = paste("PC2 (",summary_meanPCA_lm_res[2],"%)", sep=""), 
                 xlim = xlims,
                 ylim = ylims,
                 col = col.diet.means, 
                 pch = pch.gld.means, 
                 bg = col.diet.means,
                 cex = 1.5,
                 cex.axis = 1.1, 
                 cex.lab = 1.3, 
                 asp=TRUE)
text(part.coords[1], part.coords[2], "a", cex = 1.5)
text(title_coords[1],title_coords[2], "shape residual, lm-based (ecology)")

text(res_lm_PCA$x[which(rownames(res_lm_PCA$x)=="Mastacomys_fuscus"),1],-res_lm_PCA$x[which(rownames(res_lm_PCA$x)=="Mastacomys_fuscus"),2], expression(italic( "M. fuscus")), pos=1)

text(res_lm_PCA$x[which(rownames(res_lm_PCA$x)=="Conilurus_penicillatus"),1],-res_lm_PCA$x[which(rownames(res_lm_PCA$x)=="Conilurus_penicillatus"),2], expression(italic( "C. penicill.")), pos=3)

text(res_lm_PCA$x[which(rownames(res_lm_PCA$x)=="Notomys_mitchellii"),1],-res_lm_PCA$x[which(rownames(res_lm_PCA$x)=="Notomys_mitchellii"),2], expression(italic( "Notomys")), pos=3)

legend(-0.1, -0.0, legend = c("Arboreal", "Hopping", "Terrestrial", "Semiaquatic", "", "Carnivore", "Frugivore", "Granivore", "Folivore", "Omnivore"), col = c(rep("black", 4), "white", col.diet.5), border = NULL, pch = c(24:21, rep(16, 6)), cex = 0.8, bty = "n")


dev.off()




```


### Save intermediate data
```{r}
save(mean.PCA, res.PCA, res_lm_PCA, file = "../Data/Processed/04-PCA_data.rda")
```
