==> Analysis/07-heatmap-both-datasets_minimalist_plots.Rmd <==
---
title: "05-heatmap-both-datasets"
author: "Ariel Marcy"
date: "9/18/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../eco-rodents')
```

# Heatmaps both with-allometry and without-allometry 3D coordinates
1) Make heatmaps with shape dataset including allometry.
        A) Smallest to mean to largest
        B) Smallest projected shape to mean to largest projected shape
        C) PC1 minimum to mean to PC maximum

Produces multipanel Figure 2.

2) Make heatmaps for residuals aka size-less shape dataset.
        A) PC1 min to mean to PC1 max
        B) PC1 min on subset without Notomys to mean to PC1 max of this subset
        C) PC2 min to mean to PC2 max

Produces multipanel Figure 3.

### Load packages and data
Including installing libraries for landvR package, which creates the heatmaps.
```{r message = FALSE}
library(geomorph) # CRAN v4.0.6
library(devtools) # CRAN v2.4.5
if(!require(dispRity)) install.packages("dispRity")  
if(!require(landvR)) install_github("TGuillerme/landvR")  
library(landvR)   # [github::TGuillerme/landvR] v0.5.2
library(png)      # CRAN v0.1-8 
library(Rvcg)
# This sets the wd to local
library(rstudioapi) # Safely Access the RStudio API
setwd(dirname(getActiveDocumentContext()$path))

load(file = "../Data/Processed/03-main-data.rda")
load(file = "../Data/Processed/03-tree-data.rda")
load(file = "../Data/Processed/04-PCA_data.rda")
load(file = "../Data/Processed/allometry_residuals.rda")
```

## 1) Create heatmap figures
This script is dominated by the code needed to generate, export, and clean the heatmap images which are tiled in Figures 2 and 3. 

### Make heatmaps involving the smallest and largest native species 
We specify native species because technically *Mus musculus* has smallest centroid size but since it is not native, it is less relevant than *P. delicatulus*, which is only slightly larger than it.
```{r}

# Get index for smallest and largest native species 
sm.index <- which(info.means$Species == "del")
big.index <- which(info.means$MeanCsize == max(info.means$MeanCsize))

# Designate mean shapes for individual species with indices
smallest.mean <- mean.shapes[, , sm.index]
biggest.mean <- mean.shapes[, , big.index]
#which species is that?
dimnames(mean.shapes)[[3]][ big.index]


# Measure spherical coordinates differences from the smallest to the biggest mean. This results in a list of just 1 (diff.mean[[1]])
diff.mean_smallest_largest <- coordinates.difference(coordinates = biggest.mean, reference = smallest.mean, type = "spherical")


```

### Export the heatmaps for individual allometric shape changes; create "Heatmaps" folder here when running first
Export images for plots
```{r}
# Lateral view of biggest to smallest
open3d(windowRect = c(0, 0, 2400,2400))
procrustes.var.plot( smallest.mean,biggest.mean, col = heat.colors, col.val = diff.mean_smallest_largest[[1]] [, "radius"])
view3d(phi = 90)
rgl.snapshot("../Data/Processed/Heatmaps/smallest_to_biggest_meanshape.png")


```

## Make heatmaps involving smallest and largest fitted allometric shapes
First find mean specimen, then smallest and largest. Uses Dr Thomas Guillerme's  package, `landvR` to see how landmarks vary between these specimens.
```{r}
# Fit specimens
allom.fit <- procD.lm(mean.shapes ~ log(info.means$MeanCsize))
fit.shapes <- arrayspecs(allom.fit$fitted, 325, 3)

# Find largest and smallest native species from mean centroid size
smallest.shape <- fit.shapes[, , sm.index]
biggest.shape <- fit.shapes[, , big.index]

# Measure spherical coordinates differences from the mean for each specimen for each landmark
allom.diff.mean <- coordinates.difference(coordinates = biggest.shape, reference = smallest.shape, type = "spherical")


```

### Export the variation for fitted allometry shape change
Export images for plots
```{r}
# Lateral view of smallest to mean
open3d(windowRect = c(0, 0, 2400,2400))
procrustes.var.plot(smallest.shape,biggest.shape, col = heat.colors, col.val = allom.diff.mean[[1]][,"radius"])
view3d(phi = 90)
rgl.snapshot("../Data/Processed/Heatmaps/ALL_sm_big_lat.png")


```

### Make heatmaps for PC1 with allometry
Using hypothetical shapes for min and max
```{r}
# Find mean configuration - i.e. mean shape in dataset
allom.mean <- (select.procrustes(fit.shapes, selector = base::mean))[[1]]

# Ordinate the data
twoD.shape <- two.d.array(mean.shapes)
ordination <- stats::prcomp(twoD.shape)

# Force the symmetric component of shape into class "gpagen" (required for variation.range)
gpagen.shape <- list()
gpagen.shape$coords <- mean.shapes
gpagen.shape$consensus <- allom.mean
class(gpagen.shape) <- "gpagen"

# Measure extremes of variation from mean on PC1
PC1.var <- variation.range(gpagen.shape, return.ID = FALSE, axis = 1, ordination = ordination, type = "spherical")

# Wrap specimens on the tangent space
wrap.PCA <- gm.prcomp(mean.shapes)

# Select extremes (the hypotheticals that make the warp-meshes)
allom.PC1.min.h <- wrap.PCA$shapes[[1]]$max  # switched b/c axes are flipped
allom.PC1.max.h <- wrap.PCA$shapes[[1]]$min
```

### Export the PC1 variation for with allometry shape change. VW I am going from min to max here because that is the same landmark displacement direction as in the other two plots
Export images for plots
```{r}
# Lateral view of mean to smallest
open3d(windowRect = c(0, 0, 2400,2400))
procrustes.var.plot(allom.PC1.min.h, allom.PC1.max.h, col = heat.colors, col.val = PC1.var[, "radius"])
view3d(phi = 90)
rgl.snapshot("../Data/Processed/Heatmaps/PC1lg_-sm_lat.png")


```



### Make heatmaps for PC1 with residual coordinates
Steps as above.
```{r}
# Find mean configuration - i.e. mean shape in dataset
res.mean <- (select.procrustes(residuals, selector = base::mean))[[1]]

# Measure spherical coordinates differences from the mean for each specimen for each landmark
res.diff.mean <- coordinates.difference(coordinates = residuals, reference = res.mean, type = "spherical")

# Ordinate the data
twoD.shape.r <- two.d.array(residuals)
ordination.r <- stats::prcomp(twoD.shape.r)

# Force the symmetric component of shape into class "gpagen" (required for variation.range)
gpagen.shape <- list()
gpagen.shape$coords <- residuals
gpagen.shape$consensus <- res.mean
class(gpagen.shape) <- "gpagen"

# Measure extremes of variation from mean on PC1. VW I added "coords" here because, as per Thomas' explanation in the help file, this should give us the differences between the specimen with the minimum and maximun variation and gpagen.shape has $consensus which we don't want.
PC1.var.r <- variation.range(gpagen.shape$coords, return.ID = FALSE, axis = 1, ordination = ordination.r, type = "spherical")

# Wrap specimens on the tangent space
wrap.PCA.r <- gm.prcomp(residuals)

# Select extreme specimens (the ones that make the warp-meshes)
res.PC1.min.h <- wrap.PCA.r$shapes[[1]]$min
res.PC1.max.h <- wrap.PCA.r$shapes[[1]]$max
```

### Export the variation for shape residual change
```{r}
# Lateral view of mean to smallest
open3d(windowRect = c(0, 0, 2400,2400))
procrustes.var.plot(res.PC1.max.h,res.PC1.min.h, col = heat.colors, col.val = PC1.var.r[, "radius"])
view3d(phi = 90)
rgl.snapshot("../Data/Processed/Heatmaps/RES_Pc1max_min.png")


```

### Make heatmaps for PC2 residual coordinates
Steps as above but for PC2; hypothetical shapes
```{r}
# Measure extremes of variation from mean on PC1
PC1.var.r2 <- variation.range(gpagen.shape, return.ID = FALSE, axis = 2, ordination = ordination.r, type = "spherical")

# Select extreme specimens (the ones that make the warp-meshes)
res.PC2.min.h <- wrap.PCA.r$shapes[[2]]$max  # switch for consistency
res.PC2.max.h <- wrap.PCA.r$shapes[[2]]$min

# Calculate distances between shapes to ensure correct point color
min.coords <- cbind(res.mean, res.PC2.min.h)
min.dist <- apply(min.coords, 1, function(x) dist(matrix(x, nrow = 2, byrow = TRUE)))
max.coords <- cbind(res.mean, res.PC2.max.h)
max.dist <- apply(max.coords, 1, function(x) dist(matrix(x, nrow = 2, byrow = TRUE)))
```

### Export the variation for shape residual change
```{r}
# Lateral view of mean to smallest
open3d(windowRect = c(0, 0, 2400,2400))
procrustes.var.plot(res.PC2.min.h, res.PC2.max.h, col = heat.colors, col.val = min.dist)
view3d(phi = 90)
rgl.snapshot("../Data/Processed/Heatmaps/Res_PC2_min-max_lat.png")


```

### Remove Notomys and find hypothetical shapes
```{r}
# Remove Notomys from dataset
index.not <- which(info.means$Genus == "Not")
#Check check
dimnames(residuals)[[3]]==info.means$FullName

residuals.no <- residuals[, , -index.not]
info.means.no <- info.means[-index.not, ]

# Find new mean
res.mean.no <- (select.procrustes(residuals.no, selector = base::mean))[[1]]

# Measure spherical coordinates differences from the mean for each specimen for each landmark
res.diff.mean.no <- coordinates.difference(coordinates = residuals.no, reference = res.mean.no, type = "spherical")

# Ordinate the data
twoD.shape.r.n <- two.d.array(residuals.no)
ordination.r.n <- stats::prcomp(twoD.shape.r.n)

# Force the symmetric component of shape into class "gpagen" (required for variation.range)
gpagen.shape.no <- list()
gpagen.shape.no$coords <- residuals.no
gpagen.shape.no$consensus <- res.mean.no
class(gpagen.shape.no) <- "gpagen"


# Wrap specimens on the tangent space
wrap.PCA.r.n <- gm.prcomp(residuals.no)

```


### Find no Notomys PC2 shapes 
```{r}
# Select extreme specimens (the ones that make the warp-meshes)
res.PC2.min.h.n2 <- wrap.PCA.r.n$shapes[[2]]$min
res.PC2.max.h.n2 <- wrap.PCA.r.n$shapes[[2]]$max

# Calculate distances between shapes to ensure correct point color
min.coords.n2 <- cbind(res.mean.no, res.PC2.min.h.n2)
min.dist.n2 <- apply(min.coords.n2, 1, function(x) dist(matrix(x, nrow = 2, byrow = TRUE)))
max.coords.n2 <- cbind(res.mean.no, res.PC2.max.h.n2)
max.dist.n2 <- apply(max.coords.n2, 1, function(x) dist(matrix(x, nrow = 2, byrow = TRUE)))
```

### Export residual no Notomys heatmaps
```{r}
# Lateral view of mean to smallest
open3d(windowRect = c(0, 0, 2400,2400))
procrustes.var.plot(res.PC2.max.h.n2, res.PC2.min.h.n2, col = heat.colors, col.val = min.dist.n2)
view3d(phi = 90)
rgl.snapshot("../Data/Processed/Heatmaps/NO2_max-min_lat.png")

```

### Load in all heatmaps
```{r}
# Remove white space from all heatmaps (requires ImageMagik installed on PC; when installing ImageMagick, make sure you tick the option of installing legacy utilities, otherwise the below own't work)
system("mogrify ../Data/Processed/Heatmaps/*.png -trim ../Data/Processed/Heatmaps/*.png")
```

```{r}
# Load in images
Fitted_allometry <- readPNG("../Data/Processed/Heatmaps/ALL_sm_big_lat.png") 
Smalles_largest_mshape <- readPNG("../Data/Processed/Heatmaps/smallest_to_biggest_meanshape.png")
Pc1 <- readPNG("../Data/Processed/Heatmaps/PC1lg_-sm_lat.png")
Res_PC1 <- readPNG("../Data/Processed/Heatmaps/RES_Pc1max_min.png")
Res_PC2 <- readPNG("../Data/Processed/Heatmaps/Res_PC2_min-max_lat.png")
Res_PC2_no_notomys <- readPNG("../Data/Processed/Heatmaps/NO2_max-min_lat.png")

```

## 2) Make multi-panel Figure 2, heatmaps with allometry
```{r}
# Set up multi-panel figure; lateral 
#The resolution at which the raster images are displayed depends on the width and height measurements of the png file, as well as its resolution. These have to be high otherwise the image is grainy

png("../Data/Results/Figure2_Heatmap_Allometry_6panel.png", width = 2500, height = 1100, res=1400)
mat <- matrix(c(1:6), 2, 3, byrow = TRUE)  # 6plots, 3 rows, 2 columns
layout(mat)

#setting some overarching font parameters

cex_legends <- 0.25
cex_letters <- 0.3

## Fitted allometry
# 1) Large to Small, with allometry, lateral view
par(mar = c(0, 0, 0, 0) )  # sets the margins
plot(c(0, dim(Fitted_allometry)[2]), c(0, dim(Fitted_allometry)[1]), type = "n", axes = FALSE, xlab = "", ylab = "", asp = TRUE) 
rasterImage(Fitted_allometry, 0, 0, dim(Fitted_allometry)[2], dim(Fitted_allometry)[1])
text(0, dim(Fitted_allometry)[1]-10, "a", cex=cex_letters)
legend("topleft", inset = c(0.28, 0), legend = "Allometric prediction", bty = "n", cex = cex_legends)
legend("bottom", legend = "small to large", bty = "n", inset = c(0.03, 0.07), cex = cex_legends)


## Individuals
# 5) Large to Small, with allometry, lateral view
par(mar = c(0, 0, 0, 0))  # sets the margins
plot(c(0, dim(Smalles_largest_mshape)[2]), c(0, dim(Smalles_largest_mshape)[1]), type = "n", axes = FALSE, xlab = "", ylab = "", asp = TRUE) 
rasterImage(Smalles_largest_mshape, 0, 0, dim(Smalles_largest_mshape)[2], dim(Smalles_largest_mshape)[1])
text(0, dim(Smalles_largest_mshape)[1]-10, "b", cex=cex_letters)
legend("top", inset = c(0, 0), legend = "    Smallest/largest species", bty = "n", cex = cex_legends)
legend("bottom", legend = expression(paste(italic("P. delicatulus"), " to ", italic("U. caudimaculatus"))), bty = "n", inset = c(0.1, 0.07), cex = cex_legends)


## PC1 
# 9) Min to max, PCA allometry, lateral view
par(mar = c(0, 0, 0, 0))  # sets the margins
plot(c(0, dim(Pc1)[2]), c(0, dim(Pc1)[1]), type = "n", axes = FALSE, xlab = "", ylab = "", asp = TRUE) 
rasterImage(Pc1, 0, 0, dim(Pc1)[2], dim(Pc1)[1])
text(0, dim(Pc1)[1]-10, "c", cex=cex_letters)
legend("topright", inset = c(0.16, 0), legend = "Principal Component 1", bty = "n", cex = cex_legends)
legend("bottomleft", "Hypothetical maximum to minimum", bty = "n", inset = c(0.1, 0.07), cex = cex_legends)

## PC1 Residuals
# 1) Highest to lowest, residuals PC1, lateral view
par(mar = c(0, 0, 0, 0))
plot(c(0, dim(Res_PC1)[2]), c(0, dim(Res_PC1)[1]), type = "n", axes = FALSE, xlab = "", ylab = "", asp = TRUE) 
rasterImage(Res_PC1, 0, 0, dim(Res_PC1)[2], dim(Res_PC1)[1])
text(0, dim(Res_PC1)[1] - 10, "d", cex=cex_letters)
legend("topright", inset = c(0.3, 0), legend = "PC1 shape residual", bty = "n", cex = cex_legends)
legend("bottomleft", "Hypothetical maximum to minimum", bty = "n", inset = c(0.1, 0.07), cex = cex_legends)


## PC2 Residuals
# 5) Mean to Small, residuals PC2, lateral view
par(mar = c(0, 0, 0, 0))
plot(c(0, dim(Res_PC2)[2]), c(0, dim(Res_PC2)[1]), type = "n", axes = FALSE, xlab = "", ylab = "", asp = TRUE) 
rasterImage(Res_PC2, 0, 0, dim(Res_PC2)[2], dim(Res_PC2)[1])
text(0, dim(Res_PC2)[1]-10, "e", cex=cex_letters)
legend("topright", inset = c(0.3, 0), legend = "PC2 shape residual", bty = "n", cex = cex_legends)
legend("bottomleft", "Hypothetical minimum to maximum", bty = "n", inset = c(0.07, 0.07), cex = cex_legends)

# 9) Mean to Small, residuals PC2, lateral view - without Notomys
par(mar = c(0, 0, 0, 0))
plot(c(0, dim(Res_PC2_no_notomys)[2]), c(0, dim(Res_PC2_no_notomys)[1]), type = "n", axes = FALSE, xlab = "", ylab = "", asp = TRUE) 
rasterImage(Res_PC2_no_notomys, 0, 0, dim(Res_PC2_no_notomys)[2], dim(Res_PC2_no_notomys)[1])
text(0, dim(Res_PC2_no_notomys)[1]-10, "f", cex=cex_letters)
legend("topright", inset = c(0.12, 0), legend = expression(paste("PC2 residual (no ", italic("Notomys"), ")")), bty = "n", cex = cex_legends)
legend("bottomleft", "Hypothetical maximum to minimum", bty = "n", inset = c(0.1, 0.07), cex = cex_legends)



dev.off()
```

#because the allometry of the rostrum is so much lower than expected, here is a check if the heatmaps are correct - plotting shape variation "the conventional way"

```{r}
summary(allom.fit)

get_predictors <- plot(allom.fit, predictor =info.means$MeanCsize, type="regression", reg.type="RegScore")

plotRefToTarget(allom.fit$GM$fitted[,,which(dimnames(allom.fit$GM$fitted)[[3]]=="Pseudomys_delicatulus")], allom.fit$GM$fitted[,,which(dimnames(allom.fit$GM$fitted)[[3]]=="Uromys_caudimaculatus")], 
                method="vector")



```

#making a figure of meshes
```{r}

#grab stls and convert if needed
# setwd("../Data/Raw/Meshes/")
# 
# filelist <- list.files(path=".", pattern = "*stl")
# 
# for (i in 1: length(filelist)){
#   
#   convert <- vcgPlyRead(filelist[i])
#   
#   vcgPlyWrite(convert, filename = sub(".stl", ".ply", filelist[i]))
#   
# }

#grab whole path name to read in and file name for names
filelist <- list.files("../Data/Raw/Meshes/converted", full.names = TRUE)
namelist <- list.files("../Data/Raw/Meshes/converted", full.names = FALSE)
namelist <- sub("*.ply", "", namelist)
#I am here making a vector taht matches the file list, but note the scope for human error because this is manual
#namelist <- c("Hydromys chrysogaster (QM J10055)", "Mastacomys fuscus (AM M7170)", "Notomys fuscus (AM M4860)", "Pseudomys delicatulus (QM JM18715", "Uromys caudimaculatus (MV C6982)")

#read in and name meshes

for (i in 1: length(filelist)){
  
  read <- vcgPlyRead(filelist[i])
  
  assign(namelist[i], read)
  
}

#Set orientations for these meshes just once
# shade3d(Pseudomys_delicatulus_JM18715, col="pink")
# usrMat_Pseudomys <- par3d()$userMatrix # set by hand, adjust specimen i
#shade3d(Uromys_caudimaculatus_C6982, col="pink")
#usrMat_Uromys <- par3d()$userMatrix # set by hand, adjust specimen i
# shade3d(Notomys_fuscus_AM_M4860, col="pink")
# usrMat_Notomys <- par3d()$userMatrix
# shade3d(Mastacomys_fuscus_AM_M7170, col="pink")
# usrMat_Mastacomys <- par3d()$userMatrix
# shade3d(Hydromys_chrysogaster_QM_J10055, col="pink")
# usrMat_Hydromys <- par3d()$userMatrix
# 
# 
#save(usrMat_Pseudomys, usrMat_Mastacomys, usrMat_Notomys, usrMat_Uromys, usrMat_Hydromys, file="../Data/Processed/User_matrices_RGL")

load( file="../Data/Processed/User_matrices_RGL")
```

Smallest vs largest
```{r}
open3d(windowRect=c(0,0,1500, 800), zoom=0.85, FOV=0)
mfrow3d(1,2)

shade3d(Pseudomys_delicatulus_JM18715, col="pink",tran = par3d(userMatrix=usrMat_Pseudomys ), zoom=0.8 )


next3d()
shade3d(Uromys_caudimaculatus_C6982, col="pink", tran = par3d(userMatrix=usrMat_Uromys ))

#rgl.snapshot(filename = "../Data/Results/Pseudomys_vs_Uromys_meshes.png")

```

Three specialists

```{r}

cexes=1

open3d(windowRect=c(0,0,1500,800), zoom=0.85, FOV=0)
mfrow3d(1,3)


shade3d(Hydromys_chrysogaster_QM_J10055, col="pink", tran = par3d(userMatrix=usrMat_Hydromys ))
#texts3d(-10,-4,-20,texts=expression(paste(italic("H. chrysogaster"), " QM J10055")) ,cex=cexes)

next3d()
shade3d(Notomys_fuscus_AM_M4860, col="pink",tran = par3d(userMatrix=usrMat_Notomys ) )
#texts3d(-21,15,-28,texts=expression(paste(italic("N. fuscus"),  " AM M4860")), cex=cexes)

next3d()
shade3d(Mastacomys_fuscus_AM_M7170, col="pink", tran = par3d(userMatrix=usrMat_Mastacomys ))
#texts3d(14,-2,-30,texts=expression(paste(italic("M. fuscus"), " AM M7170")), cex=cexes)






rgl.snapshot(filename = "../Data/Results/Specialists.png")


```
